+++ package.json
@@ -5,17 +5,20 @@
   "dependencies": {
     "@mui/icons-material": "^5.6.2",
     "@mui/material": "^5.6.2",
+    "@mui/x-date-pickers": "^5.0.0-alpha.2",
     "@testing-library/jest-dom": "^5.16.4",
     "@testing-library/react": "^13.1.1",
     "@testing-library/user-event": "^13.5.0",
     "@types/mapbox__mapbox-gl-geocoder": "^4.7.2",
     "axios": "^0.26.1",
+    "date-fns": "^2.28.0",
     "framer-motion": "^6.3.2",
     "mapbox-gl": "^2.8.2",
     "moment": "^2.29.3",
     "moment-timezone": "^0.5.34",
     "react": "^18.0.0",
     "react-calendar": "^3.7.0",
+    "react-day-picker": "^8.0.4",
     "react-dom": "^18.0.0",
     "react-map-gl": "^7.0.11",
     "react-moment": "^1.1.2",
